+++ src/main/java/nl/inventid/rfidReader2keyboard/Read.java
@@ -21,23 +21,33 @@ import org.apache.commons.codec.binary.Hex;
 
 /**
  * This little program handles the reading of the RFID miFare chips.
+ *
+ * An exit code of 0 means everything went well
+ * An exit code of 1 means no suitable terminal was found
+ * An exit code of 2 means no type robot could be started
  */
 public class Read {
 
+	// Error messages
 	private static final String NO_CONNECT = "connect() failed";
 	private static final String EMPTY_CODE = "Scanned code was empty";
 	private static final String NO_CARD = "sun.security.smartcardio.PCSCException: SCARD_E_NO_SMARTCARD";
-	private static final String READER_UNAVAILABLE = "sun.security.smartcardio.PCSCException: SCARD_E_READER_UNAVAILABLE";
+	private static final String READER_UNAVAILABLE =
+			"sun.security.smartcardio.PCSCException: SCARD_E_READER_UNAVAILABLE";
 	private static final String FAILED_CARD_TRANSACTION =
 			"sun.security.smartcardio.PCSCException: SCARD_E_NOT_TRANSACTED";
+
+	// APDU commands
 	private static final CommandAPDU READ_COMMAND = new CommandAPDU(new byte[] { (byte) 0xFF, (byte) 0xCA, (byte) 0x00,
 			(byte) 0x00, (byte) 0x00 });
 
+	// Actual Java stuff
 	private static final List<String> TERMINAL_PREFERENCES = new ArrayList<>();
 	private static final Map<String, Integer> errorMap = new HashMap();
 	private static final String CARD_READ_FAILURE = "Card read failure";
-	public static TerminalDetector detectorLoop;
 
+	// Threads
+	public static TerminalDetector detectorLoop;
 	private static ErrorLogger errorLogger;
 
 	public static void main(String[] args) {
@@ -57,10 +67,6 @@ public class Read {
 		detectorLoop = new TerminalDetector();
 		(new Thread(detectorLoop)).start();
 
-		Read reader = new Read();
-		reader.startTerminalLoop();
-		reader.loop();
-
 		Runtime.getRuntime().addShutdownHook(new Thread() {
 			public void run() {
 				detectorLoop.stop();
@@ -69,6 +75,9 @@ public class Read {
 				System.out.println("inventid RFID capturing is now inactive. You can close this dialog");
 			}
 		});
+
+		Read reader = new Read();
+		reader.loop();
 	}
 
 	/**
@@ -78,9 +87,8 @@ public class Read {
 	 */
 	public static List<CardTerminal> listTerminals() {
 		// show the list of available terminals
-		TerminalFactory factory = TerminalFactory.getDefault();
 		try {
-			return factory.terminals().list();
+			return TerminalFactory.getDefault().terminals().list();
 		}
 		catch (Exception e) {
 			return Lists.newArrayList();
@@ -91,21 +99,11 @@ public class Read {
 		TERMINAL_PREFERENCES.add("ACS ACR122U PICC Interface"); // Best match
 		TERMINAL_PREFERENCES.add("ACR122"); // That'll do (Windows does not include the U)
 		TERMINAL_PREFERENCES.add(""); // Fuck, attach with anything (SHOULD BE LAST)
-
 	}
 
-
-	public void startTerminalLoop() {
-
-	}
-
-	/*
-	 * In face the constructor does all the work.
-	 * First a keyboardrobot is initialized
-	 * The card terminal is connected to
-	 * Then a loop is initiated which loops indefinitely (with bounds on run-away)
-	 * The loop "types" the relevant data to the PC itself
-	 * Webapps handle the rest
+	/**
+	 * Find and connect to a terminal, based on the preferences in TERMINAL_PREFERENCES
+	 * @return a valid CardTerminal or halts the system if no match can be found
 	 */
 	private CardTerminal findAndConnectToTerminal() {
 		try {
@@ -132,16 +130,33 @@ public class Read {
 			System.err.println("Unable to connect to RFID reader");
 			e.printStackTrace();
 		}
+		System.exit(1);
 		return null;
 	}
 
-	private void loop() {
+	/**
+	 * Do the actual work of the program by looping over it and writing/exceptioning In case you are looking at this
+	 * code and thinking "OMG why not just use the CardTerminal methods instead of catching?": There is a very good
+	 * reason: javax.smartcardio is buggy as fuck, so on some platform the `waitForCardPresent` and `waitForCardAbsent`
+	 * methods will not block, or block indefinitely under some conditions. Especially in combination with sleeping
+	 * code, this is a significant nightmare! Therefore we simply try to read from the card, and handle all exceptions.
+	 * In the exception handling, possibly we will reconnect to a terminal, if that is the best thing to do for
+	 * stability
+	 */
+	public void loop() {
 		CardTerminal terminal = findAndConnectToTerminal();
 		if (terminal == null) {
 			System.err.println("No terminal connected, loop is exiting");
-//			return;
 		}
-		Keyboard keyboard = new Keyboard();
+		Keyboard keyboard = null;
+		try {
+			keyboard = new Keyboard();
+		}
+		catch (Exception e) {
+			System.err.println("Could not start type robot");
+			e.printStackTrace();
+			System.exit(2);
+		}
 
 		// Random String; no UID of any chip. Still true though
 		String oldUID = "inventid bravo!";
@@ -160,41 +175,53 @@ public class Read {
 				if (!isNewCard(uid, oldUID, lastScan)) {
 					continue;
 				}
-
 				System.out.println("This is a new card! " + uid);
-				// Emulate a keyboard and "type" the uid, followed by a newline
-				keyboard.type(uid);
-				keyboard.type("\n");
 
-				i++;
+				// Emulate a keyboard and "type" the uid, followed by a newline
+				keyboard.type(uid + "\n");
+
 				oldUID = uid;
 				lastScan = Instant.now();
-				System.out.println("ready for next card");
+				i++;
 				card.disconnect(false);
+
+				System.out.println("ready for next card");
 				System.out.println("Test run: " + i);
 			}
 			catch (CardException e) {
-				// Something went wrong when scanning the card
+				// Something went wrong when scanning the card. We dont like this one, but we can easily recover from
+				// it by reconnecting. The card will have to be rescanned, but it should work then
 				if (e.getMessage().equals(FAILED_CARD_TRANSACTION) || e.getMessage().equals(READER_UNAVAILABLE)) {
 					logError(e.getMessage());
 					terminal = findAndConnectToTerminal();
 					continue;
 				}
-				// Card is not present while scanning
+				// Card is not present while scanning. This happens all the time if no card is present while the loop
+				// executes and is therefore some kind of "normal" bahaviour. Note that Windows and OSX raise this
+				// differently. Windows will raise the CardNotPresentException, OSX and Linux will throw a regular
+				// CardException.    Thanks Oracle!
 				if (e.getMessage().equals(NO_CARD) || e instanceof CardNotPresentException) {
 					logError(e.getMessage());
 					continue;
 				}
-				// Could not reliably connect to the reader (this can mean there is simply no card)
+				// Here the reader cannot connect to the card, which most often means the card was removed while
+				// reading
+				// Not a biggy. The card can also be placed while reading or moved to fast. Since the loop executed
+				// fast enough, we'll get it right on the next try
 				if (e.getMessage().equals(NO_CONNECT) || e.getMessage().equals(CARD_READ_FAILURE)) {
 					logError(e.getMessage());
 					continue;
 				}
+				// For strange readers the terminal returned an empty code. Since we do not have such cards, that is an
+				// error. We do not type the code, and log an error
 				if (e.getMessage().equals(EMPTY_CODE)) {
 					logError(e.getMessage());
 					System.err.println("Empty code was read");
 					continue;
 				}
+				// We definitely do not want to get here, since that means we triggered an error we did not expect.
+				// Therefore log everything very clearly, and for security (since we are unsure about the state, we
+				// reconnect to the reader
 				System.err.println("Help something uncatched happened! This should not happen!");
 				logError(e.getMessage());
 				e.printStackTrace();
@@ -209,7 +236,7 @@ public class Read {
 	 *
 	 * @param channel the channel to transmit over
 	 * @return a String with the value of the uid (not empty)
-	 * @throws CardException in case of an error
+	 * @throws CardException in case of an error by the terminal or if an invalid code was read
 	 */
 	private String getCardUid(CardChannel channel) throws CardException {
 		ResponseAPDU response = channel.transmit(READ_COMMAND);
@@ -234,22 +261,31 @@ public class Read {
 				(lastScan != null && lastScan.plus(1, ChronoUnit.SECONDS).isBefore(Instant.now()));
 	}
 
+	/**
+	 * Log an error count for the error
+	 *
+	 * @param errorCause the error of which the count should be incremented
+	 */
 	private void logError(String errorCause) {
 		Integer newValue = errorMap.getOrDefault(errorCause, 0) + 1;
 		errorMap.put(errorCause, newValue);
 	}
 
+	/**
+	 * This is a very stupid innerclass, which simply prints the errorMap of the main class every 60 seconds
+	 */
 	private static class ErrorLogger implements Runnable {
 
 		private boolean interrupted;
 
 		public void run() {
-			while(!interrupted) {
+			while (!interrupted) {
 				printErrorMap();
 				try {
 					Thread.sleep(60000);
 				}
 				catch (InterruptedException e) {
+					stop();
 				}
 			}
 		}
@@ -260,6 +296,30 @@ public class Read {
 
 		public void printErrorMap() {
 			System.out.println("Error map: " + errorMap.entrySet());
+		}
+	}
+
+	/**
+	 * This is a very stupid innerclass, which simply prints the connected readers every 30 seconds
+	 */
+	private static class TerminalDetector implements Runnable {
+
+		private boolean interrupted;
+
+		public void run() {
+			while (!interrupted) {
+				System.out.println(Read.listTerminals());
+				try {
+					Thread.sleep(30000);
+				}
+				catch (InterruptedException e) {
+					stop();
+				}
+			}
+		}
+
+		public void stop() {
+			interrupted = true;
 		}
 	}
